fix: add componentes
